# api-spec.yaml
openapi: 3.1.0
info:
  title: sdwanlite control-plane API (P0)
  version: 0.1.0
  description: |
    flexiWAN-inspired control plane for sdwanlite (P0 scaffold).
    All endpoints bind to loopback by default; non-loopback requires
    `--enable-live-actions` on the controller.

    Authentication: `Authorization: Bearer <bootstrap_token>` (read from
    `--bootstrap-token-file` 0600 in production, per AGENTS.md).
    `/healthz` and `/metrics` are unauthenticated and intended for the
    internal trust zone.

    Multi-tenancy: every resource carries `org_id` (UUID) and `site_id`
    (UUID). The controller MUST scope every query by `org_id`; the agent
    MUST refuse configs addressed to a different org than the one it
    registered with.

    Address hygiene: every example below uses RFC 5737 documentation
    addresses (`192.0.2.x`, `198.51.100.x`, `203.0.113.x`). No real IPs.
  contact:
    name: sdwanlite maintainers
servers:
  - url: http://127.0.0.1:8080
    description: In-process controller (loopback default)
  - url: http://controller.sdwanlite.example
    description: Production deployment (mTLS + non-loopback)
    variables: {}
paths:
  /healthz:
    get:
      operationId: healthz
      summary: Liveness probe (no auth)
      responses:
        '200':
          description: always OK
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /metrics:
    get:
      operationId: metrics
      summary: Prometheus exposition (no auth; internal scrape zone)
      responses:
        '200':
          description: text/plain; version=0.0.4
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP sdwan_controller_devices Registered device count.
                  # TYPE sdwan_controller_devices gauge
                  sdwan_controller_devices 0
  /api/v1/devices/register:
    post:
      operationId: registerDevice
      summary: Bootstrap a device (idempotent; 409 on re-register)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
            example:
              device_id: 11111111-1111-1111-1111-111111111111
              org_id:    22222222-2222-2222-2222-222222222222
              site_id:   33333333-3333-3333-3333-333333333333
              hostname:  edge-01
              version:   1
      responses:
        '201':
          description: registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RegisterResponse'}
              example:
                device_id: 11111111-1111-1111-1111-111111111111
                org_id:    22222222-2222-2222-2222-222222222222
                site_id:   33333333-3333-3333-3333-333333333333
                current_version: 1
                stream_url: ws://127.0.0.1:8080/stream/config
        '401': {description: missing or invalid bearer token}
        '409': {description: device already registered}
  /api/v1/devices/{id}/config:
    get:
      operationId: getDeviceConfig
      summary: Agent pulls the latest committed config (version-gated)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: current committed config
          content:
            application/json:
              schema: {$ref: '#/components/schemas/DeviceConfig'}
        '401': {description: missing or invalid bearer token}
        '404': {description: device not registered}
  /api/v1/devices/{id}/apply:
    post:
      operationId: applyDeviceConfig
      summary: Controller pushes a new config (transactional; agent verifies before commit)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApplyRequest'}
      responses:
        '200':
          description: applied
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApplyResponse'}
        '401': {description: missing or invalid bearer token}
        '403':
          description: org_id mismatch
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorBody'}
        '409':
          description: stale config version OR verify failure
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorBody'}
  /stream/config:
    get:
      operationId: streamConfigWs
      summary: |
        WebSocket upgrade. Server pushes the latest `DeviceConfig` whenever
        the controller commits a new version for any device. The agent sends
        the bearer token as the first text frame after the upgrade.
      security:
        - bearerAuth: []
      responses:
        '101':
          description: switching protocols — upgrade to WebSocket
  /api/v1/telemetry:
    post:
      operationId: postTelemetry
      summary: Agent pushes periodic telemetry frames (every 10s)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TelemetryFrame'}
      responses:
        '200':
          description: accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  accepted: {type: boolean}
        '401': {description: missing or invalid bearer token}
        '403':
          description: org_id mismatch
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ErrorBody'}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    RegisterRequest:
      type: object
      required: [device_id, org_id, site_id, hostname]
      properties:
        device_id: {type: string, format: uuid}
        org_id:    {type: string, format: uuid}
        site_id:   {type: string, format: uuid}
        hostname:  {type: string, minLength: 1, maxLength: 253}
        version:   {type: integer, minimum: 1, default: 1}
    RegisterResponse:
      type: object
      required: [device_id, org_id, site_id, current_version, stream_url]
      properties:
        device_id:       {type: string, format: uuid}
        org_id:          {type: string, format: uuid}
        site_id:         {type: string, format: uuid}
        current_version: {type: integer}
        stream_url:      {type: string, format: uri}
    ApplyRequest:
      type: object
      required: [config]
      properties:
        config: {$ref: '#/components/schemas/DeviceConfig'}
    ApplyResponse:
      type: object
      required: [device_id, applied_version, verified]
      properties:
        device_id:       {type: string, format: uuid}
        applied_version: {type: integer}
        verified:        {type: boolean}
    DeviceConfig:
      type: object
      required:
        - device_id
        - org_id
        - site_id
        - hostname
        - interfaces
        - tunnels
        - routes
        - firewall
        - qos
        - path_labels
        - version
      properties:
        device_id:   {type: string, format: uuid}
        org_id:      {type: string, format: uuid}
        site_id:     {type: string, format: uuid}
        hostname:    {type: string}
        interfaces:
          type: array
          items: {$ref: '#/components/schemas/Interface'}
        tunnels:
          type: array
          items: {$ref: '#/components/schemas/TunnelConfig'}
        routes:
          type: array
          items: {$ref: '#/components/schemas/Route'}
        firewall: {$ref: '#/components/schemas/FirewallPolicy'}
        qos:      {$ref: '#/components/schemas/QosPolicy'}
        path_labels:
          type: array
          items: {$ref: '#/components/schemas/PathLabel'}
        version:   {type: integer, minimum: 1}
    Interface:
      type: object
      required: [name, addresses]
      properties:
        name: {type: string, example: eth0}
        addresses:
          type: array
          items: {type: string, format: ipv4}
          example: ['203.0.113.7']
        mtu: {type: integer, default: 0}
        path_label: {type: string, nullable: true}
    TunnelConfig:
      oneOf:
        - $ref: '#/components/schemas/WireGuardTunnel'
      discriminator:
        propertyName: kind
    WireGuardTunnel:
      type: object
      required: [kind, interface, path_label, endpoint, allowed_ips, public_key]
      properties:
        kind: {type: string, enum: [wire_guard]}
        interface: {type: string, example: wg0}
        path_label: {type: string}
        health_check: {$ref: '#/components/schemas/HealthCheckConfig'}
        endpoint: {type: string, example: '203.0.113.7:51820'}
        allowed_ips:
          type: array
          items: {type: string, format: ipv4}
        public_key: {type: string, description: 'base64 X25519 public key (44 chars)'}
    HealthCheckConfig:
      type: object
      required: [interval_ms, probe_type, threshold, timeout_ms]
      properties:
        interval_ms: {type: integer, default: 1000}
        probe_type:
          type: string
          enum: [icmp, http, dns, tcp]
        threshold: {type: integer, default: 3}
        timeout_ms: {type: integer, default: 500}
    Route:
      type: object
      required: [destination, next_hop]
      properties:
        destination: {type: string, example: '0.0.0.0/0'}
        next_hop:    {type: string, example: '203.0.113.1'}
        metric:      {type: integer, default: 100}
    FirewallPolicy:
      type: object
      properties:
        rules:
          type: array
          items: {$ref: '#/components/schemas/FirewallRule'}
    FirewallRule:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, drop, reject]
        source:      {type: string, nullable: true}
        destination: {type: string, nullable: true}
        protocol:    {type: string, nullable: true}
        port:        {type: integer, nullable: true}
        comment:     {type: string, nullable: true}
    QosPolicy:
      type: object
      properties:
        classes:
          type: array
          items: {$ref: '#/components/schemas/QosClass'}
    QosClass:
      type: object
      required: [name, dscp]
      properties:
        name: {type: string}
        dscp: {type: integer, minimum: 0, maximum: 63}
        bandwidth_bps: {type: integer, default: 0}
    PathLabel:
      type: object
      required: [id, name, type, sla]
      properties:
        id:   {type: string, format: uuid}
        name: {type: string}
        type:
          type: string
          enum: [mpls, internet, 5g, starlink, lte, other]
        sla:  {type: string}
    TelemetryFrame:
      type: object
      required: [device_id, org_id, uptime_secs, links, flags]
      properties:
        device_id:   {type: string, format: uuid}
        org_id:      {type: string, format: uuid}
        uptime_secs: {type: integer}
        links:
          type: array
          items: {$ref: '#/components/schemas/LinkSample'}
        flags:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/HealthFlagLinkDown'
              - $ref: '#/components/schemas/HealthFlagDegraded'
    LinkSample:
      type: object
      required: [path_label, interface, local_endpoint, tx_bytes, rx_bytes]
      properties:
        path_label:    {type: string}
        interface:     {type: string}
        local_endpoint: {type: string, example: '203.0.113.7:51820'}
        tx_bytes:      {type: integer}
        rx_bytes:      {type: integer}
        peer_endpoint: {type: string, nullable: true}
    HealthFlagLinkDown:
      type: object
      required: [kind, path_label]
      properties:
        kind: {type: string, enum: [link_down]}
        path_label: {type: string}
    HealthFlagDegraded:
      type: object
      required: [kind, subsystem]
      properties:
        kind: {type: string, enum: [degraded]}
        subsystem: {type: string}
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:   {type: string, description: 'machine-readable code, NEVER echoes the underlying error'}
        message: {type: string, description: 'see server logs'}